third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGAnimatedInteger-expected.txt

This is a testharness.js-based test.
[FAIL] SVGAnimatedInteger interface - utilizing the targetX property of SVGFEConvolveMatrix
  assert_equals: expected 3 but got -1
Harness: the test ran to completion.
